requirements.txt
pymupdf==1.26.6
imagehash>=4.3.0
Pillow>=10.0.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
